Handle setting the connection's cursor factory
The `cursor_factory` attribute of a connection object can be set to
define cursor behaviour (e.g. return records as dicts or named
tuples). Previously a `default_cursor_factory` set a default
behaviour*, but changing this wouldn't propagate to a connection that
has already been established (i.e. when the value of the `db` property
is referenced).

This switches the default cursor factory from an attribute to a
property so a side effect on setting the value can be added to update
the connection at the same time.

*default behaviour because the intent is to support overriding the
cursor factory when performing a query.