Add an esipy adapter/wrapper.
Functionality for making requests to the EVE ESI API is now
implemented via `esipy`. This is kept de-coupled via an ESIClient
which provides some convenience methods and behaviour to keep keep a
clean simple interface.

Currently EsiClient and EsiApp instances are cached on a ESIClient
per-instance basis. Whether this is the right or wrong thing to do
remains to be discovered...