Add MarketOrder model with tests
The MarketOrder class models a market buy or sell order and attempts
to be source-agnostic. Fields that are used in this codebase are
exposed from the underlying data dictionary as properties, with any
type casting implemented within the getter as appropriate to
introduce useful standardisation.

A data dictionary is expected on construction, but an additional
constructor is provided for creating the MarketOrder directly from a
JSON string.