Add a Market class for fetching timestamped orders
This class acts as a MarketOrderSnapshot store, organising orders by
region/system/location and allocating a timestamp reflecting they time
they were queried.

There is a mild issue here in that the update method doesn't know if
cache is being used, so the timestamp may well update but the data may
not. This isn't expected to be a big issue.