Add models for regions, systems and stations
These are built from an EVE static data export by reference to their
type and ID or name. Because they are read-only, they are memoised for
efficiency and nice semantics ("if location is other_location").

Note that the source data is copied into instance attributes at the
moment; it might not be necessary (there are a finite, not-ridiculous
number of locations and we don't use much info) but we could just
reference the dicts returned and cached on the EveStaticData instance
and wrap them rather than duplicate fields. There might be
implications to consider here about multiple EveStaticData instances
which is avoidable but possible in the current design.